+++ src/main/java/ai/deepcode/jbplugin/DeepCodeStartupActivity.java
@@ -26,7 +26,7 @@ public class DeepCodeStartupActivity implements StartupActivity {
     // Keep commented - for DEBUG ONLY !!!!!!!!!!!!!!!!!
     //PropertiesComponent.getInstance(project).setValue("consentGiven", false);
 
-    AnalysisData.clearCache(project);
+    AnalysisData.resetCachesAndTasks(project);
     // Initial logging if needed.
     if (LoginUtils.isLogged(project, true)) {
       RunUtils.asyncAnalyseProjectAndUpdatePanel(project);