+++ src/main/java/ai/deepcode/jbplugin/core/RunUtils.java
@@ -21,42 +21,6 @@ import java.util.stream.Collectors;
 public class RunUtils {
   private RunUtils() {}
 
-  private static final Map<Project, Integer> mapProject2RequestsCounter = new ConcurrentHashMap<>();
-
-  private static int getBulkRequestsCount(@NotNull Project project) {
-    return mapProject2RequestsCounter.computeIfAbsent(project, p -> 0);
-  }
-
-  /** No events for individual files should be processed */
-  public static boolean inBulkMode(@NotNull Project project) {
-    return getBulkRequestsCount(project) > 0;
-  }
-
-  public static void setBulkMode(@NotNull Project project) {
-    final int counter = getBulkRequestsCount(project) + 1;
-    if (counter == 1) {
-      // cancel all running tasks first
-      cancelRunningIndicators(project);
-    }
-    DCLogger.info("BulkMode ON with " + counter + " total requests");
-    mapProject2RequestsCounter.put(project, counter);
-  }
-
-  public static void unsetBulkMode(@NotNull Project project) {
-    final int counter = getBulkRequestsCount(project) - 1;
-    if (counter >= 0) {
-      mapProject2RequestsCounter.put(project, counter);
-      DCLogger.info("BulkMode OFF with " + counter + " total requests");
-    } else {
-      DCLogger.warn("BulkMode OFF request with already " + counter + " total requests");
-    }
-  }
-
-  public static void forceUnsetBulkMode(@NotNull Project project) {
-    mapProject2RequestsCounter.put(project, 0);
-    DCLogger.info("BulkMode OFF forced");
-  }
-
   public static void asyncUpdateCurrentFilePanel(PsiFile psiFile) {}
 
   public static <T> T computeInReadActionInSmartMode(
@@ -110,6 +74,7 @@ public class RunUtils {
     @Override
     public void run(@NotNull ProgressIndicator indicator) {
       DCLogger.info("New Process started at " + project);
+      indicator.setIndeterminate(false);
       getRunningIndicators(project).add(indicator);
 
       runnable.run();
@@ -138,7 +103,7 @@ public class RunUtils {
             .collect(Collectors.joining("\n"));
     DCLogger.info("Canceling ProgressIndicators:\n" + indicatorsList);
     // in case any indicator holds Bulk mode process
-    forceUnsetBulkMode(project);
+    BulkMode.forceUnset(project);
     getRunningIndicators(project).forEach(ProgressIndicator::cancel);
     getRunningIndicators(project).clear();
     projectsWithFullRescanRequested.remove(project);
@@ -172,6 +137,7 @@ public class RunUtils {
             new Task.Backgroundable(project, "DeepCode: Analysing Files...") {
               @Override
               public void run(@NotNull ProgressIndicator indicator) {
+                indicator.setIndeterminate(false);
 
                 // To let new event cancel the currently running one
                 ProgressIndicator prevProgressIndicator =
@@ -188,6 +154,7 @@ public class RunUtils {
                           + "]");
                   prevProgressIndicator.cancel();
                   getRunningIndicators(project).remove(prevProgressIndicator);
+                  HashContentUtils.removeHashContent(psiFile);
                 }
                 getRunningIndicators(project).add(indicator);
 
@@ -240,7 +207,7 @@ public class RunUtils {
     if (inBulkMode) bulkModeRequests.add(requestId);
     if (prevRequestId != null) {
       if (bulkModeRequests.remove(prevRequestId)) {
-        RunUtils.unsetBulkMode(project);
+        BulkMode.unset(project);
       }
       return;
     }
@@ -255,6 +222,7 @@ public class RunUtils {
             new Task.Backgroundable(project, "DeepCode: Analysing Files...") {
               @Override
               public void run(@NotNull ProgressIndicator indicator) {
+                indicator.setIndeterminate(false);
 
                 // To let new event cancel the currently running one
                 ProgressIndicator prevProgressIndicator =
@@ -276,7 +244,7 @@ public class RunUtils {
                 // unset BulkMode if cancelled process did run under BulkMode
                 Long prevRequestId = mapProject2CancellableRequestId.put(project, requestId);
                 if (prevRequestId != null && bulkModeRequests.remove(prevRequestId)) {
-                  RunUtils.unsetBulkMode(project);
+                  BulkMode.unset(project);
                 }
 
                 // delay to let new consequent requests proceed and cancel current one
@@ -293,11 +261,11 @@ public class RunUtils {
                   mapProject2RequestId.remove(project);
 
                   // actual rescan
-                  AnalysisData.removeProjectFromCache(project);
+                  AnalysisData.removeProjectFromCaches(project);
                   updateCachedAnalysisResults(project, null);
 
                   if (bulkModeRequests.remove(actualRequestId)) {
-                    RunUtils.unsetBulkMode(project);
+                    BulkMode.unset(project);
                   }
                 } else {
                   DCLogger.warn("No actual RequestId found for: " + project.getName());